chore(Handlers):create authentication handlers
- create Route Handlers folder
- separate routes from their handlers
- added signup handlers
- added login handlers
- added password reset handler
- added articles handlers
- added comments handlers
- added like and dislike handlers
[Finishes #172393890]
